% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{castillo2024.rgmomentum.e2}
\alias{castillo2024.rgmomentum.e2}
\title{Data from Experiment 2 in \insertCite{castillo2024ExplainingFlawsHuman;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 28483 rows and 20 columns.
}
\source{
\url{https://osf.io/dw8ez/}
}
\usage{
castillo2024.rgmomentum.e2
}
\description{
Participants first learned a set of syllables arranged in either a single row (one-dimensional condition) or a grid (two-dimensional condition), then produced two random sequences for the same display.
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/dw8ez/}{OSF}.
\describe{
\item{id}{participant id}
\item{part_Gender}{participant's gender (self-reported)}
\item{part_Age}{participant's age (self-reported)}
\item{index}{position of the item in the sequence, 0 indexed}
\item{id}{unique identifier for the participant}
\item{block}{whether the item belongs to the first sequence the participant uttered (A) or the second (B)}
\item{syll}{syllable uttered}
\item{starts}{timestamp of when the utterance starts, in seconds.}
\item{delays}{temporal difference with the start of the previous item (i.e. \code{starts[index] - starts[index - 1]})}
\item{dim}{whether the participant was allocated to the one-dimensional or two-dimensional condition}
\item{seed}{Which of five possible configurations the participant learned}
\item{position}{The position of the syllable in the array. For 1D arrays, position is left to right. For 2D arrays positions 1-2 correspond to the top 2 cells; 3-5 to the middle 3 cells; and 6-7 to the bottom three cells (always left to right)}
\item{R}{whether the item is a repetition of the last}
\item{A}{whether the item is adjacent to the last in the display (after removing repetitions)}
\item{TP_full}{whether the item is a turning point, considering all items (after removing repetitions)}
\item{D}{the Euclidean distance to the previous item (after removing repetitions)}
\item{S}{a measure of how likely the item is in a uniform or gaussian distribution (see text)}
\item{expected_*}{the expectation for measure \code{*} derived from reshuffling the participant's sequence 10000 times}
}
}
\references{
\insertRef{castillo2024ExplainingFlawsHuman}{samplrData}
}
\keyword{datasets}
