% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplrData-package.R
\docType{package}
\name{samplrData-package}
\alias{samplrData}
\alias{samplrData-package}
\title{samplrData: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucas-castillo/samplrData}
  \item \url{https://lucas-castillo.github.io/samplrData/}
  \item Report bugs at \url{https://github.com/lucas-castillo/samplrData/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
