% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spicer2022.anchoringrepulsion.e1}
\alias{spicer2022.anchoringrepulsion.e1}
\title{Data from Experiment 1 in \insertCite{spicer2022PerceptualCognitiveJudgments;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 9600 rows and 11 columns.
}
\source{
\url{https://osf.io/95ruy/}
}
\usage{
spicer2022.anchoringrepulsion.e1
}
\description{
Perceptual judgments. Participants made judgments of numerosity against comparison values or absolute estimates. Comparison values (boundaries) were either similar or dissimilar to the true answer.
}
\details{
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/95ruy/}{OSF}.
\describe{
\item{Timestamp}{Date and time of the experimental session}
\item{Pt}{Participant ID}
\item{Trial}{Trial ID based on order of presentation}
\item{Boundary}{Comparison value for that trial}
\item{DotCount}{Number of dots shown on that trial}
\item{Region}{Region for that dot count, being either high or low}
\item{Decision}{Decision made by the participant on whether dot count was higher or lower than the boundary for that trial}
\item{Dec_RT}{Response time for the decision}
\item{Accuracy}{Accuracy of the selected decision}
\item{Estimate}{Direct estimate of the number of dots on that trial made by the participant. NaN is used for trials in which no estimate was requested}
\item{Est_RT}{Response time for the estimate}
}
}
\references{
\insertRef{spicer2022PerceptualCognitiveJudgments}{samplrData}
}
\keyword{datasets}
