% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spicer2022.anchoringrepulsion.e2}
\alias{spicer2022.anchoringrepulsion.e2}
\title{Data from Experiment 2 in \insertCite{spicer2022PerceptualCognitiveJudgments;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 2960 rows and 13 columns.
}
\source{
\url{https://osf.io/95ruy/}
}
\usage{
spicer2022.anchoringrepulsion.e2
}
\description{
Cognitive judgments. Participants answered questions about commonly experienced values. judgments of numerosity against comparison values or absolute estimates. Comparison values (boundaries) were either similar or dissimilar to the true answer.
}
\details{
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/95ruy/}{OSF}.
\describe{
\item{Timestamp}{Date and time of the experimental session}
\item{Pt}{Participant ID}
\item{Trial}{Trial ID based on order of presentation}
\item{QID}{ID for the target question of that trial}
\item{Question}{Question text}
\item{Region}{Expected region for that question, being either high or low}
\item{Answer}{Unbiased answer for that question from calibration data}
\item{Boundary}{Comparison value for that trial}
\item{Decision}{Decision made by the participant on whether answer to the question was higher or lower than the boundary}
\item{Dec_RT}{Response time for the decision}
\item{Accuracy}{Accuracy of the selected decision based on calibration data}
\item{Estimate}{Direct estimate of the answer to the question for that trial made by the participant}
\item{Est_RT}{Response time for the estimate}
}
}
\references{
\insertRef{spicer2022PerceptualCognitiveJudgments}{samplrData}
}
\keyword{datasets}
