% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sundh2023.meanvariance.e4}
\alias{sundh2023.meanvariance.e4}
\title{Data from Experiment 4 in \insertCite{sundh2023UnifiedExplanationVariability;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 13320 rows and 7 columns.
}
\source{
\url{https://osf.io/9kea6/}
}
\usage{
sundh2023.meanvariance.e4
}
\description{
Participants made probability judgments about future hypothetical events, of the format: “What is the probability that there will be an early UK general election AND the UK economy will recover this year?". The experiment consisted of three blocks, so that all participants responded to each unique query three times.
}
\details{
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/9kea6/}{OSF}.
\describe{
\item{ID}{}
\item{block}{3 blocks in total}
\item{query, querydetail}{Verbal descriptions of the query}
\item{querytype}{Type of query: e.g. not B given A = p(¬B|A)}
\item{queryset}{Whether the query is about biden and 2050 climate goals or UK election and economic recovery}
\item{Estimate}{Estimated probability, in percentages}
}
}
\references{
\insertRef{sundh2023UnifiedExplanationVariability}{samplrData}
}
\keyword{datasets}
