% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2022.coherenceaccuracy.e1}
\alias{zhu2022.coherenceaccuracy.e1}
\title{Data from Experiment 1 in \insertCite{zhu2022ClarifyingRelationshipCoherence;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 82 rows and 23 columns.
}
\source{
\url{https://osf.io/cdvkn/}
}
\usage{
zhu2022.coherenceaccuracy.e1
}
\description{
Participants (from Prolific) estimated the frequencies of different 3-card combinations in a 52 card deck and 3-ball combinations in a 52 ball urn (mathematically identical questions). They also answered surveys on poker playing habits and gamblers fallacy questionnaire.
}
\details{
See exact questions in original paper's supplementary materials (Appendix B). These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/cdvkn/}{OSF}.
\describe{
\item{group}{Self-reported response on whether they have played poker before}
\item{q1-q9}{Answers to the poker questions}
\item{mq1-mq9}{Answers to the ball questions}
\item{gfs}{number of correct answers in gambler's fallacy questionnaire}
\item{cs}{Inferred poker playing time in the last 12 months}
\item{RT}{}
\item{taskEqual}{judged similarity between the Card and Ball task (0=all equal, 1=all differ, 0.5=answers differ but urn and deck were equal)}
}
}
\references{
\insertRef{zhu2022ClarifyingRelationshipCoherence}{samplrData}
}
\keyword{datasets}
