% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2022.structurenoise.animals}
\alias{zhu2022.structurenoise.animals}
\title{Data from Animal Experiment in \insertCite{zhu2022UnderstandingStructureCognitive;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 4967 rows and 7 columns.
}
\source{
\url{https://osf.io/kcfgp/}
}
\usage{
zhu2022.structurenoise.animals
}
\description{
Participants were asked to type animal names as they came to mind and were explicitly instructed that they could resubmit previous animals, though not consecutively.
}
\details{
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/kcfgp/}{OSF}.
\describe{
\item{ID}{Participant ID}
\item{Order}{Index of the response}
\item{Responses}{Transcribed response}
\item{Animal}{Category the response was allocated to}
\item{StartType,EndType}{Absolute time of starting and ending to type the response}
\item{IRI}{Time between last response's EndType and this response's StartType}
}
}
\references{
\insertRef{zhu2022UnderstandingStructureCognitive}{samplrData}
}
\keyword{datasets}
