% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2022.structurenoise.time}
\alias{zhu2022.structurenoise.time}
\title{Data from Time Experiment in \insertCite{zhu2022UnderstandingStructureCognitive;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 29822 rows and 6 columns.
}
\source{
\url{https://osf.io/kcfgp/}
}
\usage{
zhu2022.structurenoise.time
}
\description{
Participants first listened to a sample of the target temporal interval for 60 seconds. Participants were instructed to reproduce the target by pressing the spacebar when they believed the target interval had elapsed (i.e. perfect performance in the task would mean \code{IRI == Target}).
}
\details{
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/kcfgp/}{OSF}.
\describe{
\item{ID}{Participant ID}
\item{Order}{Index of the response}
\item{StartType,EndType}{Absolute time of starting and ending to type the response}
\item{IRI}{Time between last response's EndType and this response's StartType}
\item{Target}{Whether the participant had to reproduce a 1/3s, 1s or 3s interval}
}
}
\references{
\insertRef{zhu2022UnderstandingStructureCognitive}{samplrData}
}
\keyword{datasets}
