% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplrData-package.R
\docType{package}
\name{samplrData-package}
\alias{samplrData}
\alias{samplrData-package}
\title{samplrData: Datasets from the SAMPLING Project}
\description{
Contains human behaviour datasets collected by the SAMPLING project (\url{https://sampling.warwick.ac.uk}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucas-castillo/samplrData}
  \item \url{https://lucas-castillo.github.io/samplrData/}
  \item Report bugs at \url{https://github.com/lucas-castillo/samplrData/issues}
}

}
\author{
\strong{Maintainer}: Lucas Castillo \email{lucas.castillo-marti@warwick.ac.uk} (\href{https://orcid.org/0000-0003-0274-0777}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Yun-Xiao Li \email{yunxiao.li@warwick.ac.uk} (\href{https://orcid.org/0000-0002-3509-6618}{ORCID}) [copyright holder]
  \item Adam N Sanborn \email{a.n.sanborn@warwick.ac.uk} (\href{https://orcid.org/0000-0003-0442-4372}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item European Research Council (ERC) [funder]
}

}
\keyword{internal}
