% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2020.bayesiansampler.e2}
\alias{zhu2020.bayesiansampler.e2}
\title{Data from Experiment 2 in \insertCite{zhu2020BayesianSamplerGeneric;textual}{samplrData}}
\format{
An object of class \code{data.frame} with 22380 rows and 10 columns.
}
\source{
\url{https://osf.io/mgcxj/}
}
\usage{
zhu2020.bayesiansampler.e2
}
\description{
Participants made probability judgments of the format: “What is the probability that the weather is [X] on a random day in England?". Various weather events were used, and the queries included both marginal events, conditional events, conjunctions, and disjunctions. The total set of 20 unique queries formed a block within which the presentation order was randomized for each participant. The experiment consisted of three blocks, so that all participants responded to each unique query three times.
}
\details{
These data are licensed under \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}, reproduced from materials in \href{https://osf.io/mgcxj/}{OSF}.
\describe{
\item{ID}{}
\item{block}{3 blocks in total}
\item{trial}{Trial Number within a block}
\item{query, querydetail}{Verbal descriptions of the query}
\item{querytype}{Type of query: e.g. notBgA = p(¬B|A)}
\item{Estimate}{Estimated probability, in percentages}
\item{starttime, endtime}{}
\item{RT}{}
}
}
\references{
\insertRef{zhu2020BayesianSamplerGeneric}{samplrData}
}
\keyword{datasets}
